Create SCF response function module and update relevant modules
