+++ pyscf/solvent/_attach_solvernt.py
